requirements.txt
arrow==0.14.7
astroid==2.4.2
beautifulsoup4==4.9.1
bs4==0.0.1
colorama==0.4.3
html5lib==1.1
ics==0.7
isort==5.4.2
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
pylint==2.6.0
python-dateutil==2.8.1
six==1.15.0
soupsieve==2.0.1
TatSu==5.5.0
toml==0.10.1
webencodings==0.5.1
wrapt==1.12.1
